+++ src/views/users/show.vue
@@ -90,8 +90,8 @@
                   <td>{{ $t('userProfile.status') }}</td>
                   <td>
                     <el-tag v-if="!user.is_approved" type="info">{{ $t('userProfile.pending') }}</el-tag>
-                    <el-tag v-if="!user.deactivated && user.is_approved" type="success">{{ $t('userProfile.active') }}</el-tag>
-                    <el-tag v-if="user.deactivated" type="danger">{{ $t('userProfile.deactivated') }}</el-tag>
+                    <el-tag v-if="user.is_active && user.is_approved" type="success">{{ $t('userProfile.active') }}</el-tag>
+                    <el-tag v-if="!user.is_active" type="danger">{{ $t('userProfile.deactivated') }}</el-tag>
                   </td>
                 </tr>
               </tbody>